backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
praw==7.7.1
httpx==0.25.2
python-multipart==0.0.6
openai==1.12.0
